+++ Scripts/UI/UIFader.cs
@@ -68,6 +68,15 @@ namespace AeLa.Utilities.UI
 		/// </summary>
 		public Coroutine FadeInCoroutine()
 		{
+			if (!isActiveAndEnabled)
+			{
+				// can't start coroutine so force set state
+				OnFadeInStart.Invoke();
+				canvasGroup.alpha = FadeInCurve.Evaluate(1f);
+				OnFadeInComplete.Invoke();
+				return null;
+			}
+
 			StopAllCoroutines();
 			return StartCoroutine(FadeRoutine(1));
 		}
@@ -77,6 +86,15 @@ namespace AeLa.Utilities.UI
 		/// </summary>
 		public Coroutine FadeOutCoroutine()
 		{
+			if (!isActiveAndEnabled)
+			{
+				// can't start coroutine so force set state
+				OnFadeOutStart.Invoke();
+				canvasGroup.alpha = FadeOutCurve.Evaluate(1f);
+				OnFadeOutComplete.Invoke();
+				return null;
+			}
+
 			StopAllCoroutines();
 			return StartCoroutine(FadeRoutine(-1));
 		}